requirements.txt
albumentations==1.3.1
ansi2html==1.8.0
anyio==3.7.1
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
attrs==23.1.0
audioread==3.0.0
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
cachetools==5.3.1
catboost==1.1.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.5
cmake==3.26.4
comm==0.1.3
contourpy==1.0.7
cycler==0.11.0
dash==2.11.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
exceptiongroup==1.1.2
executing==1.2.0
fastjsonschema==2.17.1
filelock==3.9.0
Flask==2.2.5
fonttools==4.39.3
fqdn==1.5.1
google==3.0.0
google-api-core==2.11.1
google-api-python-client==2.92.0
google-auth==2.21.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.59.1
graphviz==0.20.1
greenlet==2.0.2
grpcio==1.56.0
grpcio-status==1.56.0
httplib2==0.22.0
idna==3.4
imageio==2.31.1
ipykernel==6.24.0
ipython==8.14.0
ipython-genutils==0.2.0
ipywidgets==8.0.7
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.18.0
jsonschema-specifications==2023.6.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyter_server==2.7.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.8
kiwisolver==1.4.4
lazy_loader==0.3
librosa==0.10.0.post2
lightning-utilities==0.9.0
lit==16.0.6
llvmlite==0.40.1
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mistune==3.0.1
mpmath==1.2.1
msgpack==1.0.5
nbclassic==1.0.0
nbclient==0.8.0
nbconvert==7.6.0
nbformat==5.9.1
nest-asyncio==1.5.6
networkx==3.0
notebook==6.5.4
notebook_shim==0.2.3
numba==0.57.1
numpy==1.24.2
opencv-python==4.8.0.74
opencv-python-headless==4.8.0.74
overrides==7.3.1
packaging==23.0
pandas==2.0.0
pandasql==0.7.3
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.8.1
plotly==5.14.1
pooch==1.6.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
protobuf==4.23.4
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
py4j==0.10.9.7
pyarrow==12.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycocotools==2.0.6
pycparser==2.21
pydub==0.25.1
Pygments==2.15.1
pyparsing==3.0.9
pyspark==3.4.1
python-dateutil==2.8.2
python-json-logger==2.0.7
python-speech-features==0.6
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==25.1.0
qudida==0.0.4
referencing==0.29.1
requests==2.28.1
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.8.10
rsa==4.9
scikit-image==0.21.0
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
Send2Trash==1.8.2
six==1.16.0
sniffio==1.3.0
soundfile==0.12.1
soupsieve==2.4.1
soxr==0.3.5
SQLAlchemy==2.0.18
stack-data==0.6.2
sympy==1.11.1
tenacity==8.2.2
terminado==0.17.1
threadpoolctl==3.1.0
tifffile==2023.7.4
tinycss2==1.2.1
torch==2.0.0+cu117
torch-summary==1.4.5
torchaudio==2.0.1+cu117
torchinfo==1.7.2
torchmetrics==1.0.1
torchvision==0.15.1+cu117
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
triton==2.0.0
typing_extensions==4.4.0
tzdata==2023.3
uri-template==1.3.0
uritemplate==4.1.1
urllib3==1.26.13
visdom==0.2.4
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.1
Werkzeug==2.2.3
widgetsnbextension==4.0.8
